Move pkg/subctl/operator/lighthouse to pkg/lighthouse
