+++ pkg/lighthouse/scc/ensure.go
